Fixed ambiguity

+++ examples/ArtnetnodeWifiSendDmx/ArtnetnodeWifiSendDmx.ino
@@ -54,11 +54,11 @@ void onDmxFrame(uint16_t universe, uint16_t length, uint8_t sequence, uint8_t* d
 {
   // Send "break" as a "slow" zero.
   Serial.begin(56700);
-  Serial.write(0);
+  Serial.write((uint8_t)0);
   delayMicroseconds(220);
   Serial.begin(250000, SERIAL_8N2);
   
-  Serial.write(0); // Start-Byte
+  Serial.write((uint8_t)0); // Start-Byte
   // send out the buffer
   for (int i = 0; i < length; i++)
   {
